view-task & institute registration api calls implemented; major code …

+++ server/controllers/instituteController.js
@@ -1,7 +1,7 @@
-const asyncHandler = require('express-async-handler');
-const User = require('../models/userModel');
-const Institute = require('../models/instituteModel');
-const Task = require('../models/taskModel');
+import asyncHandler from "express-async-handler";
+import User from "../models/userModel.js";
+import Institute from "../models/instituteModel.js";
+import Task from "../models/taskModel.js";
 
 const getInstitutes = asyncHandler(async (req, res) => {
   const institutes = await Institute.find({});
@@ -9,9 +9,11 @@ const getInstitutes = asyncHandler(async (req, res) => {
 });
 
 const getInstituteById = asyncHandler(async (req, res) => {
-  const { name } = req.body;
+  // const { name } = req.body;
+  const { name } = req.query;
+
   if (!name) {
-    return res.status(400).json({ message: 'Institute name is required' });
+    return res.status(400).json({ message: "Institute name is required" });
   }
   const institute = await Institute.findOne({
     name,
@@ -19,51 +21,49 @@ const getInstituteById = asyncHandler(async (req, res) => {
   if (institute) {
     res.json(institute);
   } else {
-    return res.status(404).json({ message: 'Institute not found' });
+    return res.status(404).json({ message: "Institute not found" });
   }
 });
 
 const createInstitute = asyncHandler(async (req, res) => {
-  const { name, nodalOfficerId, logo, reportingOfficer } = req.body;
-  if (!name || !nodalOfficerId) {
-    return res
-      .status(400)
-      .json({ message: 'Name and Nodal officer ID are required' });
+  const { name, nodalOfficerId, logo, reportingOfficerId } = req.body;
+
+  if (!name || !nodalOfficerId || !reportingOfficerId) {
+    return res.status(400).json({
+      message:
+        "Name of Institute, Nodal officer ID, Reporting officer ID are required",
+    });
   }
   const instituteExists = await Institute.findOne({
     name,
   });
+
   if (instituteExists) {
-    return res.status(400).json({ message: 'Institute already exists' });
+    return res.status(400).json({ message: "Institute already exists" });
   }
+
   const nodalOfficer = await User.findOne({
     userId: nodalOfficerId,
-    role: 'nodal',
   });
 
   if (!nodalOfficer) {
-    return res.status(400).json({ message: 'Nodal officer not found' });
+    return res.status(400).json({ message: "Nodal officer not found" });
   }
 
-  if (reportingOfficer) {
-    const foundReportingOfficer = await User.findOne({
-      userId: reportingOfficer,
-      role: 'reporting',
-    });
+  const reportingOfficer = await User.findOne({
+    userId: reportingOfficerId,
+  });
 
-    console.log(foundReportingOfficer);
-
-    if (!foundReportingOfficer) {
-      return res.status(400).json({ message: 'Reporting officer not found' });
-    }
+  if (!reportingOfficer) {
+    return res.status(400).json({ message: "Reporting officer not found" });
   }
 
   const institute = new Institute({
     name,
     nodalOfficer: nodalOfficerId,
-    reportingOfficer: reportingOfficer || null,
+    reportingOfficer: reportingOfficerId,
     tasks: [],
-    logo: logo || null,
+    // logo: logo || null,
   });
 
   const createdInstitute = await institute.save();
@@ -74,7 +74,7 @@ const assignReportingOfficer = asyncHandler(async (req, res) => {
   const { reportingOfficerId, instituteName } = req.body;
   if (!reportingOfficerId || !instituteName) {
     return res.status(400).json({
-      message: 'Institute name and Reporting officer ID are required',
+      message: "Institute name and Reporting officer ID are required",
     });
   }
 
@@ -84,18 +84,18 @@ const assignReportingOfficer = asyncHandler(async (req, res) => {
   if (institute) {
     const reportingOfficer = await User.findOne({
       userId: reportingOfficerId,
-      role: 'reporting',
+      role: "reporting",
     });
 
     if (!reportingOfficer) {
-      return res.status(400).json({ message: 'Reporting officer not found' });
+      return res.status(400).json({ message: "Reporting officer not found" });
     }
 
     institute.reportingOfficer = reportingOfficerId;
     const updatedInstitute = await institute.save();
     return res.json(updatedInstitute);
   } else {
-    return res.status(404).json({ message: 'Institute not found' });
+    return res.status(404).json({ message: "Institute not found" });
   }
 });
 
@@ -104,7 +104,7 @@ const assignNodalOfficer = asyncHandler(async (req, res) => {
   if (!nodalOfficerId || !instituteName) {
     return res
       .status(400)
-      .json({ message: 'Institute name and Nodal officer ID are required' });
+      .json({ message: "Institute name and Nodal officer ID are required" });
   }
 
   const institute = await Institute.findOne({
@@ -113,18 +113,18 @@ const assignNodalOfficer = asyncHandler(async (req, res) => {
   if (institute) {
     const nodalOfficer = await User.findOne({
       userId: nodalOfficerId,
-      role: 'nodal',
+      role: "nodal",
     });
 
     if (!nodalOfficer) {
-      return res.status(400).json({ message: 'Nodal officer not found' });
+      return res.status(400).json({ message: "Nodal officer not found" });
     }
 
     institute.nodalOfficer = nodalOfficerId;
     const updatedInstitute = await institute.save();
     return res.json(updatedInstitute);
   } else {
-    return res.status(404).json({ message: 'Institute not found' });
+    return res.status(404).json({ message: "Institute not found" });
   }
 });
 
@@ -132,7 +132,7 @@ const addTaskToInstitute = asyncHandler(async (req, res) => {
   const { taskId, instituteName } = req.body;
   if (!taskId || !instituteName) {
     return res.status(400).json({
-      message: 'Institute name and Task ID are required',
+      message: "Institute name and Task ID are required",
     });
   }
 
@@ -145,13 +145,13 @@ const addTaskToInstitute = asyncHandler(async (req, res) => {
       taskId,
     });
     if (!task) {
-      return res.status(400).json({ message: 'Task not found' });
+      return res.status(400).json({ message: "Task not found" });
     }
     institute.tasks.push(taskId);
     const updatedInstitute = await institute.save();
     return res.json(updatedInstitute);
   } else {
-    return res.status(404).json({ message: 'Institute not found' });
+    return res.status(404).json({ message: "Institute not found" });
   }
 });
 
@@ -159,7 +159,7 @@ const updateLogo = asyncHandler(async (req, res) => {
   const { logo, instituteName } = req.body;
   if (!logo || !instituteName) {
     return res.status(400).json({
-      message: 'Institute name and Logo are required',
+      message: "Institute name and Logo are required",
     });
   }
   const institute = await Institute.findOne({
@@ -170,43 +170,66 @@ const updateLogo = asyncHandler(async (req, res) => {
     const updatedInstitute = await institute.save();
     return res.json(updatedInstitute);
   } else {
-    return res.status(404).json({ message: 'Institute not found' });
+    return res.status(404).json({ message: "Institute not found" });
+  }
+});
+
+const updateInstitute = asyncHandler(async (req, res) => {
+  const { logo, name, instituteId: _id } = req.body;
+
+  // if (!logo || !name) {
+  if (!name) {
+    return res.status(400).json({
+      message: "Institute name and Logo are required",
+    });
+  }
+  const institute = await Institute.findOne({
+    _id,
+  });
+
+  if (institute) {
+    institute.logo = logo;
+    institute.name = name;
+    const updatedInstitute = await institute.save();
+    return res.json(updatedInstitute);
+  } else {
+    return res.status(404).json({ message: "Institute not found" });
   }
 });
 
 const deleteInstitute = asyncHandler(async (req, res) => {
   const { name } = req.body;
   if (!name) {
-    return res.status(400).json({ message: 'Institute name is required' });
+    return res.status(400).json({ message: "Institute name is required" });
   }
   const institute = await Institute.findOne({
     name,
   });
-  
+
+  if (!institute) {
+    return res.status(404).json({ message: "Institute not found" });
+  }
+
   const nodalOfficer = institute.nodalOfficer;
   const reportingOfficer = institute.reportingOfficer;
-  
+
   if (nodalOfficer) {
     await User.findOneAndDelete({
       userId: nodalOfficer,
-    })
+    });
   }
 
   if (reportingOfficer) {
     await User.findOneAndDelete({
       userId: reportingOfficer,
-    })
+    });
   }
 
-  if (institute) {
-    await institute.remove();
-    return res.json({ message: 'Institute removed' });
-  } else {
-    return res.status(404).json({ message: 'Institute not found' });
-  }
+  await Institute.deleteOne({ _id: institute._id });
+  return res.json({ message: "Institute removed" });
 });
 
-module.exports = {
+export {
   getInstitutes,
   getInstituteById,
   createInstitute,
@@ -215,4 +238,5 @@ module.exports = {
   addTaskToInstitute,
   updateLogo,
   deleteInstitute,
+  updateInstitute,
 };
